backend/requirements.txt
#test for push

google-genai
pydantic
python-dotenv